% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_folder_info.R
\name{extract_folder_info}
\alias{extract_folder_info}
\title{Extract folder info}
\usage{
extract_folder_info(path, attribute)
}
\arguments{
\item{path}{The path from where you want to list folder(s)}

\item{attribute}{The attribute you want to extract. For a complete list of attributes
see `?file.info()`}
}
\value{
A vector of attributes requested
}
\description{
Extract folder info
}
\examples{
\dontrun{
extract_folder_info("tests", "ctime")
}
}
