% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_folder.R
\name{list_folder}
\alias{list_folder}
\title{List folders from a given directory.}
\usage{
list_folder(path, verbose = FALSE)
}
\arguments{
\item{path}{The path from where you want to list folder(s)}

\item{verbose}{Boolean deciding how verbose the output should be. Defaults to FALSE.}
}
\value{
Returns a dataframe of folders in the assigned path.
}
\description{
List folders from a given directory.
}
\examples{
\donttest{
list-folders("tests")
}
}
