% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narcissist.R
\name{narcissist}
\alias{narcissist}
\title{Check pkg stats on CRAN}
\usage{
narcissist(pkg, from)
}
\arguments{
\item{pkg}{The name of the package which stats you want to obtain}

\item{from}{The start date for fetched figures}
}
\value{
Returns a dataframe with results
}
\description{
Check pkg stats on CRAN
}
\examples{
\dontrun{
narcissist(pkg = "zzlite", from = "2020-02-25")
}
}
