% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_folders.R
\name{list_folders}
\alias{list_folders}
\title{List folders from a given directory.}
\usage{
list_folders(path)
}
\arguments{
\item{path}{The path from where you want to list folder(s)}
}
\value{
Returns a list of folders
}
\description{
List folders from a given directory.
}
\examples{
\donttest{
list-folders("tests")
}
}
